Add module file with evaluate function
