+++ package.json
@@ -2,7 +2,7 @@
   "name": "aatt",
   "version": "0.0.1",
   "description": "Automated Accessibility Testing Tool",
-  "main": "app.js",
+  "main": "module.js",
   "dependencies": {
     "body-parser": "~1.10.1",
     "consolidate": "~0.10.0",
